Don't override files in fetch script

+++ fetchInput.sh
@@ -1,4 +1,17 @@
 day=$1
-curl -sS --cookie "session=$(cat cookie.txt)" "https://adventofcode.com/2021/day/$day/input" -o src/input/day$day.txt
-cp src/skeleton.py src/day$day.py
-sed -i "s/'REPLACE_ME'/$day/g" src/day$day.py
+inputFile="src/input/day$day.txt"
+if [[ ! -f $inputFile ]]; then
+  curl -sS --cookie "session=$(cat cookie.txt)" "https://adventofcode.com/2021/day/$day/input" -o $inputFile
+  echo "Fetched $inputFile"
+else
+  echo "$inputFile already exists"
+fi
+
+sourceFile="src/day$day.py"
+if [[ ! -f $sourceFile ]]; then
+  cp src/skeleton.py src/day$day.py
+  sed -i "s/'REPLACE_ME'/$day/g" src/day$day.py
+  echo "Created $sourceFile"
+else
+  echo "$sourceFile already exists"
+fi
